DDF-UI-165 fixed cluster selection problem and turned popup-preview into hooks
