+++ ui-frontend/packages/catalog-ui-search/package.json
@@ -66,7 +66,6 @@
     "density-clustering": "1.3.0",
     "dropzone": "4.3.0",
     "eonasdan-bootstrap-datetimepicker": "Eonasdan/bootstrap-datetimepicker#4.17.47",
-    "fetch": "^1.1.0",
     "font-awesome": "4.7",
     "geo-convex-hull": "1.2.0",
     "geolib": "^3.2.1",
