+++ ui-frontend/packages/catalog-ui-search/src/main/webapp/react-component/popup-preview/index.tsx
@@ -13,9 +13,4 @@
  *
  **/
 
-export type Metacard = {
-  getPreview: Function
-  getTitle: Function
-}
-
 export { default } from './popup-preview'
