+++ ui-frontend/packages/catalog-ui-search/src/main/webapp/react-component/popup-preview/popup-preview.tsx
@@ -14,12 +14,95 @@
  **/
 import * as React from 'react'
 import { hot } from 'react-hot-loader'
-import withListenTo, { WithBackboneProps } from '../backbone-container'
+import styled from 'styled-components'
 
-import { Metacard } from '.'
+import { useBackbone } from '../../component/selection-checkbox/useBackbone.hook'
+
+export type Metacard = {
+  getPreview: Function
+  getTitle: Function
+}
+
+const Root = styled.div`
+  font-family: 'Inconsolata', 'Lucida Console', monospace;
+  background: ${props => props.theme.backgroundModal};
+  display: block;
+  width: auto;
+  height: auto;
+  font-size: ${props => props.theme.mediumFontSize};
+  position: absolute;
+  text-align: left;
+  padding: 4px;
+  max-height: 290px;
+  max-width: 50%;
+  transform: translate(-52.5%, -100%);
+
+  &::before {
+    top: 100%;
+    content: ' ';
+    border-top: 15px solid ${props => props.theme.backgroundModal};
+    border-left: 10px solid transparent;
+    border-right: 10px solid transparent;
+    height: 0;
+    width: 0;
+    left: 50%;
+    position: absolute;
+    pointer-events: none;
+  }
+`
+
+const Title = styled.div`
+  font-size: 20px;
+  margin: 0;
+  padding: 2px 6px;
+  white-space: nowrap;
+  overflow: hidden;
+  text-overflow: ellipsis;
+  font-family: 'Open Sans', arial, sans-serif;
+`
+
+const Preview = styled.div`
+  position: relative;
+  min-width: 200px;
+  height: 100%;
+  min-height: 15px;
+  max-height: 250px;
+  padding: 2px;
+  white-space: normal;
+  background-color: ${props => props.theme.backgroundContent};
+  border: 1px solid;
+  overflow-y: auto;
+  overflow-x: auto;
+  text-overflow: ellipsis;
+`
+
+const PreviewHtml = styled.html`
+  font-family: 'Open Sans', arial, sans-serif;
+  font-size: 14px;
+  padding: 2px 4px;
+`
+
+const ClusterList = styled.ul`
+  margin: 1px;
+  padding: 1px;
+  border: 1px solid;
+`
+
+const ClusterTitle = styled.li`
+  font-size: 18px;
+  margin: 0;
+  padding: 2px 6px;
+  white-space: nowrap;
+  overflow: hidden;
+  text-overflow: ellipsis;
+  font-family: 'Open Sans', arial, sans-serif;
+
+  &:hover {
+    background-color: ${props => props.theme.backgroundSlideout};
+  }
+`
 
 const sanitize = require('sanitize-html')
-const PopupPresentation = require('./presentation').default
 
 const NO_PREVIEW = 'No preview text available.'
 const STATUS_OK = 200
@@ -39,137 +122,114 @@ const hasPreview = (header: string) => {
   return br2nl(sanitized) !== NO_PREVIEW
 }
 
-const mapPropsToState = (props: Props) => {
-  const { map } = props
-  const metacard = map.get('popupMetacard')
-  const clusterModels = map.get('popupClusterModels')
-  const location = map.get('popupLocation')
-
-  return {
-    showPopup: location && (metacard || clusterModels),
-    left: location ? location.left + 'px' : 0,
-    top: location ? location.top - TOP_OFFSET + 'px' : 0,
-  }
-}
-
 type Props = {
   map: Backbone.Model
-} & WithBackboneProps
-
-type State = {
-  showPopup: Boolean
-  titleText?: String
-  previewHtml?: String
-  clusterModels?: Array<Metacard>
-  clusterTitleCallback?: Function
 }
 
-class PopupPreview extends React.Component<Props, State> {
-  constructor(props: Props) {
-    super(props)
-    this.state = {
-      ...mapPropsToState(props),
-      showPopup: false,
-      clusterTitleCallback: this.clusterTitleCallback,
-    }
-  }
+const getLeft = ({ map }: Props) => {
+  const location = map.get('popupLocation')
+  return location ? location.left + 'px' : 0
+}
 
-  componentDidMount() {
-    this.listenToMap()
-  }
+const getTop = ({ map }: Props) => {
+  const location = map.get('popupLocation')
+  return location ? location.top - TOP_OFFSET + 'px' : 0
+}
 
-  listenToMap = () => {
-    const { listenTo, map } = this.props
-    listenTo(map, 'change:popupLocation', this.handlePopupChange)
-    listenTo(map, 'change:popupMetacard change:popupClusterModels', () => {
-      this.setPopup(this.props)
+const getShowPopup = ({ map }: Props) => {
+  const location = map.get('popupLocation')
+  const clusterModels = map.get('popupClusterModels')
+  const metacard = map.get('popupMetacard')
+  return location && (metacard || clusterModels)
+}
+
+const getPreviewHTML = ({
+  map,
+  setPreviewHTML,
+}: {
+  map: any
+  setPreviewHTML: React.Dispatch<React.SetStateAction<string | undefined>>
+}) => {
+  if (map.get('popupMetacard')) {
+    const url = map.get('popupMetacard').getPreview() as string
+    const xhr = new XMLHttpRequest()
+    xhr.addEventListener('load', () => {
+      if (xhr.status === STATUS_OK) {
+        setPreviewHTML(
+          hasPreview(xhr.responseText) ? xhr.responseText : undefined
+        )
+      }
     })
-  }
 
-  componentWillUnmount() {
-    const { stopListening, map } = this.props
-    stopListening(map, 'change:popupLocation', this.handlePopupChange)
-    stopListening(
+    xhr.open('GET', url)
+    xhr.send()
+  } else {
+    setPreviewHTML(undefined)
+  }
+}
+
+const HookPopupPreview = (props: Props) => {
+  const { map } = props
+  const forceRender = React.useState(Math.random())[1]
+  const [previewHTML, setPreviewHTML] = React.useState(undefined as
+    | undefined
+    | string)
+  const { listenTo } = useBackbone()
+
+  React.useEffect(() => {
+    listenTo(
       map,
-      'change:popupMetacard change:popupClusterModels',
-      this.setPopup
+      'change:popupMetacard change:popupClusterModels change:popupLocation',
+      () => {
+        forceRender(Math.random())
+        getPreviewHTML({ map, setPreviewHTML })
+      }
     )
+  }, [])
+
+  if (!getShowPopup(props)) {
+    return null
   }
 
-  handlePopupChange = () => {
-    this.setState(mapPropsToState(this.props))
-  }
-
-  setPopup = (props: Props) => {
-    const metacard = props.map.get('popupMetacard')
-    const clusterModels = props.map.get('popupClusterModels')
-
-    if (metacard) {
-      this.setPopupMetacard(metacard)
-    } else if (clusterModels) {
-      this.setPopupCluster(clusterModels)
-    }
-  }
-
-  /**
-    Set state values for a single metacard
-   */
-  setPopupMetacard = (metacard: Metacard) => {
-    this.setpreviewHtml(metacard)
-    this.setState({
-      titleText: metacard.getTitle(),
-      clusterModels: undefined,
-    })
-  }
-
-  /**
-   Set state values for a cluster
-  */
-  setPopupCluster = (clusterModels: Array<Metacard>) => {
-    this.setState({ titleText: undefined, previewHtml: undefined })
-    this.setState({ clusterModels })
-  }
-
-  /**
-    Gets the previewHtml from the targetMetacard url
-   */
-  setpreviewHtml = (metacard: Metacard) => {
-    const url = metacard.getPreview()
-
-    if (url) {
-      const xhr = new XMLHttpRequest()
-      xhr.addEventListener('load', () => {
-        if (xhr.status === STATUS_OK) {
-          this.setState({
-            previewHtml: hasPreview(xhr.responseText)
-              ? xhr.responseText
-              : undefined,
-          })
+  return (
+    <Root style={{ left: getLeft(props), top: getTop(props) }}>
+      {(function() {
+        if (map.get('popupMetacard')) {
+          const metacardJSON = map.get('popupMetacard').toJSON()
+          return (
+            <>
+              <Title>{metacardJSON.metacard.properties.title}</Title>
+              {previewHTML && (
+                <Preview>
+                  <PreviewHtml
+                    dangerouslySetInnerHTML={{ __html: previewHTML }}
+                  />
+                </Preview>
+              )}
+            </>
+          )
+        } else if (map.get('popupClusterModels')) {
+          const clusterModels = map.get('popupClusterModels')
+          return (
+            <ClusterList>
+              {clusterModels.map((clusterModel: any) => {
+                return (
+                  <ClusterTitle
+                    key={clusterModel.id}
+                    onClick={() => {
+                      map.set('popupMetacard', clusterModel)
+                    }}
+                  >
+                    {clusterModel.toJSON().metacard.properties.title}
+                  </ClusterTitle>
+                )
+              })}
+            </ClusterList>
+          )
         }
-      })
-
-      xhr.open('GET', url)
-      xhr.send()
-    } else {
-      this.setState({ previewHtml: undefined })
-    }
-  }
-
-  /**
-   *  Switch from cluster to metacard when cluster titles are clicked
-   */
-  clusterTitleCallback = (title: String) => {
-    const metacard = this.state!.clusterModels!.find(
-      m => m.getTitle() === title
-    )
-    if (metacard) {
-      this.setPopupMetacard(metacard)
-    }
-  }
-
-  render() {
-    return this.state.showPopup ? <PopupPresentation {...this.state} /> : null
-  }
+      })()}
+    </Root>
+  )
 }
 
-export default hot(module)(withListenTo(PopupPreview))
+export default hot(module)(HookPopupPreview)
